+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- feat: added prettify --line N for in-file table detection and reformat
+
 ## v0.2.1 (2026-06-13)
 
 - feat: removed reddit and table-el styles